REFACTOR: src/Products/controllers
