+++ src/Products/controllers/category/categoryController.js
@@ -1,4 +1,4 @@
-const { EntityCategory } = require('../db');
+const { EntityCategory } = require('../../../db.js');
 
 
 const createCategory = async (req, res) => {
